Implemented review suggestions (https://github.com/WomenPlusPlus/deploy-impact-22-wikimini-b/issues/77)
