+++ src/node-backend/adapters/database-adapter.js
@@ -1,33 +1,6 @@
-import mariadb from "mariadb";
-import * as dotenv from 'dotenv'
-dotenv.config()
+import {pool} from "./database-connection.js";
 
-export const pool = mariadb.createPool({
-  host: process.env.HOST,
-  user: process.env.MARIA_USER,
-  password: process.env.MARIA_PASS,
-  database: process.env.MARIA_DATABASE,
-  connectionLimit: 5,
-});
-
-// Connect and check for errors
-pool.getConnection((err, connection) => {
-  if (err) {
-    if (err.code === "PROTOCOL_CONNECTION_LOST") {
-      console.error("Database connection lost");
-    }
-    if (err.code === "ER_CON_COUNT_ERROR") {
-      console.error("Database has too many connection");
-    }
-    if (err.code === "ECONNREFUSED") {
-      console.error("Database connection was refused");
-    }
-    console.error("Unknown error while initialising database connection: " + err)
-  }
-  if (connection) connection.release();
-});
-
-// add a list of student codes to the db, with teacher, class and student names given
+// not tested yet, to do when focusing on classroom
 export async function addStudentCodes(teacherUsername, className, studentCodes, studentNames) {
   try {
     // queries to be tested
@@ -52,7 +25,6 @@ export async function addStudentCodes(teacherUsername, className, studentCodes, 
 }
 
 // pass student code and verify that it exists, hasn't expired yet and give back the teacher username (and email?) and classroom
-// could be improved to give back more detail: expired/not found
 export async function verifyCode(studentCode) {
   if (studentCode.length !== 5) {
     throw Error("Student code has to be 5 characters long");
@@ -61,8 +33,8 @@ export async function verifyCode(studentCode) {
     const checkCodeQuery = "SELECT teachers.username, classrooms.name, classroomCodes.studentFullName, classroomCodes.used FROM teachers, classrooms, classroomCodes " +
         "WHERE teachers.username=classroomCodes.teacherUsername and classrooms.id=classroomCodes.classID and classroomCodes.studentJoinCode=?";
     const queryResult = await pool.query(checkCodeQuery, studentCode);
-    const valid = (queryResult.length === 1) && (queryResult[0]["used"] === false); // is valid if there is only one corrsponding code and if it hasn't been used yet
-    return {"valid": valid, "result": queryResult}; // is that good?
+    const valid = (queryResult.length === 1) && (queryResult[0]["used"] === false);
+    return {"valid": valid, "result": queryResult};
   } catch (error) {
     console.error("Error while checking student code: " + error);
     return error;
